refactor to seperate holo-crypto and wasm
