+++ src/holo-crypto/.gitignore
@@ -1,4 +1,4 @@
-/target
+target/
 **/*.rs.bk
 Cargo.lock
 bin/
