+++ src/holo-crypto/crates/holo-crypto-wasm/profiling/bootstrap.js
