+++ src/holo-crypto/crates/holo-crypto/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "holo-crypto"
+version = "0.1.0"
+authors = ["Willem Olding <willemolding@gmail.com>"]
+edition = "2018"
+
+[lib]
+crate-type = ["lib"]
+
+[dependencies]
+base64 = "0.10.1"
+rust-argon2 = "0.5.1"
+ed25519-dalek = "1.0.0-pre.2"
+wasm-bindgen = "0.2"
