+++ src/holo-crypto/crates/holo-crypto/README.md
@@ -1,0 +1,3 @@
+# Holo-Crypto
+
+Exposes key derivation and signing functionality required by HP-admin.
