+++ src/holo-crypto/crates/holo-crypto/src/error.rs
@@ -1,3 +1,4 @@
+
 use wasm_bindgen::JsValue;
 
 #[derive(Debug)]
